Restructure application folder and files
