+++ lambdatimes/src/components/Tabs/Tabs.js
