+++ lambdatimes/src/components/pages/news/News.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import TopBar from '../../layout/TopBar';
+import Header from '../../layout/Header';
+import CardContainer from '../../CardContainer/CardContainer';
+
+export const News = () => {
+	return (
+		<React.Fragment>
+			<TopBar />
+			<Header />
+			<CardContainer />
+		</React.Fragment>
+	);
+};
